Completed "Dependencies and Liaisons" section with actual list of groups
The "Dependencies" sub-section is now explicit about potential common web
technologies that the group could refer to and on the possible overlap
with the Network Service Discovery API. I also moved the bits on the
community group to that sub-section.

The "Liaisons" sub-section now lists to a supposedly appropriate set of
groups within W3C and explains the motivation behing each liaison.

I am not entirely sure about the external bodies that are to appear in
the "External Groups" sub-section. I listed possible bodies that the WG
should collaborate with to ensure that the API can be implemented on top of
a variety of widely deployed protocols. I may have missed a couple. Even if
protocols end up "out of scope" for the working group, it seems important to
keep the list around.