Refined scope: no protocols, possible guidelines, interoperability defined
Main changes:
1. protocols have been moved out of scope
2. the success criteria section mentions interoperability and attempts
to define it somehow.
3. the other deliverables section mentions the possibility to produce
implementation guidelines

More detailed list of changes:

- New final sentence to the "Goals" section. It seemed weird to finish the
section on a sentence that used the negative form without describing what the
group would do to solve that issue.

- Dropped the protocol subsection from the Scope section and reordered the
text in that section accordingly

- Added protocols to the Out of Scope section, using former text. Also added
a note about referencing existing suites of protocols in informative notes
to promote interoperability.

- Updated the text on content mirroring as it seemed awkward to present second
screens differently in that section. Plus the Presentation API is not really
meant to display content on screens that are "off one edge of the primary
screen"

- Dropped the Phase 2 subsection in the Deliverables section.

- Added the possibility to work on informative guidelines, either directly
within the Presentation API or within a separate Note.