Added PING, Web Security IG and WebRTC to liaisons list
The charter mentions privacy and security issues. As such, reviews by the
Privacy Interest Group and Web Security Interest Group are strongly
recommended.

I added the WebRTC group as well in the list of groups that develop
specifications that are potentially relevant for implementors of the
Presentation API, e.g. for extending the API to support use cases that
the group will not address per se (such as content mirroring).