Scope: completed 2UA case to mention multiple controlling pages
The paragraph in the Scope section that describes the 2UA use case was not
clear whether connecting to a running session from another authorized user
agent would disconnect the first user agent or whether multiple controlling
pages can run in parallel.

The commit adds a "at the same time" in:
  "pages hosted by other user agents may be authorized to control the
   remotely rendered content at the same time."

... to make it more explicit that the Presentation API will support multiple
controlling pages.